backend/api_server/requirements.txt
# Core FastAPI and web framework (Modal compatible versions)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# HTTP requests and APIs
requests==2.32.5
httpx==0.28.1
certifi==2024.8.30

# Authentication and security
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4

# File handling and utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Modal.com deployment
modal

# Date and time handling
python-dateutil==2.8.2

# Email validation
email-validator==2.1.0.post1
pydantic[email]

# Cryptography
cryptography==41.0.7

# Additional lightweight backend essentials
python-slugify==8.0.1
validators==0.20.0